Move users API definitions to its app
